Feature/speed up setup matching (#1067)
* Add concurrent functions for all get methods

* Replace itergets with concurrent_gets

* Fix: Get offset and limit if available

The dict.get method returns the value of the passed key. If the value
of the key is None, it will return None. Since we do not want None present
we use 'or' to get a default offset and limit value

* Feat: Convert concurrent_get to get_all

get_all_notes is easier to understand than concurrent_get_notes
get_all methods were moved to openreview.py and api.client.py
This way, it's easier to use the methods: instead of doing
openreview.tools.concurrent_get_notes, one can simply use
client.get_all_notes with the same params that get_notes
uses

* Feat: Replace all concurrent_get methdos with get_all methods

* Feat: Replace iterget_notes with get_all_notes

Co-authored-by: Melisa Bok <melisabok@gmail.com>
Co-authored-by: celestemartinez <32438984+celestemartinez@users.noreply.github.com>