Account for profile ID
The original PR wasn't accounting for the fact that when multiple accounts use the same name, numbers are used to differentiate between preferred names, so most names were not automatically passing. The tests were passing because there were no publications associated with the name, which also triggers automatic acceptance.

This PR updates the tests and the process function to account for that fact