Update request_remove_name_process.py (#2419)
* Update request_remove_name_process.py

Fix the reversed names that weren't being approved.

* Account for profile ID

The original PR wasn't accounting for the fact that when multiple accounts use the same name, numbers are used to differentiate between preferred names, so most names were not automatically passing. The tests were passing because there were no publications associated with the name, which also triggers automatic acceptance.

This PR updates the tests and the process function to account for that fact

* Change preferred name to pretty=True

Makes it cleaner so we don't need to pre-process the preferred name as much. Also updated test_remove_name_from_merged_profile so it doesn't fail anymore

---------

Co-authored-by: celestemartinez <32438984+celestemartinez@users.noreply.github.com>